:bug: Expand "matches" to generic YouTube domain & Add onUpdated Listener
This is due to a problem where YCLF doens't activate when a user enters YouTube via youtube.com and selects the video later on, rather than visiting the watch page right away. If a user visits youtube.com and moves to the watch page, chrome doesn't recognize the url change and cannot activate the filter. This commit addresses the issue by expanding the "matches" url to generic YouTube domain, and firing up an Listener that manually checks the current url is the watch page.